seperate object factory from tkobject.h
