+++ ToolKit/Entity.cpp
@@ -287,7 +287,7 @@ namespace ToolKit
       {
         int type = -1;
         ReadAttr(comNode, XmlParamterTypeAttr, type);
-        Component* com = GetComponentFactory()->Create((ComponentFactory::ComponentType) type);
+        Component* com = ComponentFactory::Create((ComponentFactory::ComponentType) type);
         com->m_version = m_version;
         com->DeSerialize(info, comNode);
         AddComponent(std::shared_ptr<Component>(com));
