+++ ToolKit/Scene.cpp
@@ -590,7 +590,7 @@ namespace ToolKit
     {
       XmlAttribute* typeAttr      = node->first_attribute(xmlObjectType);
       EntityFactory::EntityType t = (EntityFactory::EntityType) std::atoi(typeAttr->value());
-      EntityPtr ntt               = GetEntityFactory()->CreateByType(t);
+      EntityPtr ntt               = EntityFactory::CreateByType(t);
       ntt->m_version              = m_version;
 
       ntt->DeSerialize(info, node);
