+++ ToolKit/TKObject.cpp
@@ -128,92 +128,4 @@ namespace ToolKit
     ParameterEventConstructor(); // Set all the events after data deserialized.
   }
 
-  TKObjectFactory::TKObjectFactory() { Init(); }
-
-  TKObjectFactory::~TKObjectFactory() {}
-
-  TKObjectFactory::ObjectConstructorCallback& TKObjectFactory::GetConstructorFn(const StringView Class)
-  {
-    auto constructorFnIt = m_constructorFnMap.find(Class);
-    if (constructorFnIt != m_constructorFnMap.end())
-    {
-      return constructorFnIt->second;
-    }
-
-    return m_nullFn;
-  }
-
-  /**
-   * Constructs a new TKObject from class name.
-   * @param cls - Class name of the object to be created.
-   * @return A new instance of the object with the given class name.
-   */
-  TKObject* TKObjectFactory::MakeNew(const StringView Class)
-  {
-    if (auto constructorFn = GetConstructorFn(Class))
-    {
-      TKObject* object = constructorFn();
-      object->NativeConstruct();
-      return object;
-    }
-
-    assert(false && "Unknown object type.");
-    return nullptr;
-  }
-
-  void TKObjectFactory::Init()
-  {
-    // Entities.
-    Register<AABBOverrideComponent>();
-    Register<AnimControllerComponent>();
-    Register<DirectionComponent>();
-    Register<EnvironmentComponent>();
-    Register<MaterialComponent>();
-    Register<MeshComponent>();
-    Register<SkeletonComponent>();
-    Register<Arrow2d>();
-    Register<AudioSource>();
-    Register<Billboard>();
-    Register<Camera>();
-    Register<Cone>();
-    Register<Cube>();
-    Register<Drawable>();
-    Register<Entity>();
-    Register<EntityNode>();
-    Register<Light>();
-    Register<DirectionalLight>();
-    Register<PointLight>();
-    Register<SpotLight>();
-    Register<LineBatch>();
-    Register<Prefab>();
-    Register<Quad>();
-    Register<SkyBase>();
-    Register<GradientSky>();
-    Register<Sky>();
-    Register<Sphere>();
-    Register<Quad>();
-    Register<SpriteAnimation>();
-    Register<Surface>();
-    Register<Canvas>();
-    Register<Button>();
-    // Resources.
-    Register<Animation>();
-    Register<Audio>();
-    Register<Material>();
-    Register<Mesh>();
-    Register<SkinMesh>();
-    Register<Scene>();
-    Register<Shader>();
-    Register<Skeleton>();
-    Register<SpriteSheet>();
-    Register<Texture>();
-    Register<CubeMap>();
-    Register<DataTexture>();
-    Register<LightDataTexture>();
-    Register<SSAONoiseTexture>();
-    Register<DepthTexture>();
-    Register<Hdri>();
-    Register<RenderTarget>();
-  }
-
 } // namespace ToolKit
